packages/backend/requirements.txt
# Core Web Framework (Flask only - remove FastAPI/Starlette conflicts)
flask>=3.0.0
flask-cors>=4.0.0

# Database and Vector Storage  
chromadb>=0.4.0
sqlalchemy>=2.0.0
langchain-chroma>=0.1.0

# AI and LLM Dependencies
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0
langchain-text-splitters>=0.0.1
openai>=1.0.0
google-generativeai>=0.3.0
sentence-transformers>=2.2.0
transformers>=4.30.0
tokenizers>=0.13.0
huggingface-hub>=0.16.0

# Machine Learning and Data Processing (Compatible versions)
torch>=2.0.0
numpy>=1.24.0,<2.0.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.10.0
matplotlib>=3.7.0
faiss-cpu>=1.7.4

# Document Processing
pypdf>=3.0.0
PyMuPDF>=1.23.0
python-pptx>=0.6.0
openpyxl>=3.1.0
pytesseract>=0.3.10
Pillow>=9.5.0
opencv-python>=4.8.0

# Web Scraping and HTTP
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0

# Authentication and Security
bcrypt>=4.0.0
pydantic>=2.0.0

# Development and Testing
python-dotenv>=1.0.0
rich>=13.0.0
tqdm>=4.65.0
pyyaml>=6.0.0